Homework 17
updated test for add song to playlist
